+++ lib/rubocop/cop/style/indent_hash.rb
@@ -51,16 +51,40 @@ module Rubocop
           left_brace = hash_node.loc.begin
           return if first_pair.loc.expression.line == left_brace.line
 
+          if separator_style?(first_pair)
+            check_based_on_longest_key(hash_node.children, left_brace,
+                                       left_parenthesis)
+          else
+            check_first_pair(first_pair, left_brace, left_parenthesis, 0)
+          end
+        end
+
+        def separator_style?(first_pair)
+          separator = first_pair.loc.operator
+          key = "Enforced#{separator.is?(':') ? 'Colon' : 'HashRocket'}Style"
+          config.for_cop('AlignHash')[key] == 'separator'
+        end
+
+        def check_based_on_longest_key(pairs, left_brace, left_parenthesis)
+          key_lengths = pairs.map do |pair|
+            pair.children.first.loc.expression.length
+          end
+          check_first_pair(pairs.first, left_brace, left_parenthesis,
+                           key_lengths.max - key_lengths.first)
+        end
+
+        def check_first_pair(first_pair, left_brace, left_parenthesis, offset)
           column = first_pair.loc.expression.column
           @column_delta =
-            expected_column(left_brace, left_parenthesis) - column
+            expected_column(left_brace, left_parenthesis, offset) - column
 
           if @column_delta == 0
             correct_style_detected
           else
             add_offense(first_pair, :expression,
                         message(base_description(left_parenthesis))) do
-              if column == unexpected_column(left_brace, left_parenthesis)
+              if column == unexpected_column(left_brace, left_parenthesis,
+                                             offset)
                 opposite_style_detected
               else
                 unrecognized_style_detected
@@ -70,7 +94,7 @@ module Rubocop
         end
 
         # Returns the expected, or correct indentation.
-        def expected_column(left_brace, left_parenthesis)
+        def expected_column(left_brace, left_parenthesis, offset)
           base_column =
             if left_parenthesis && style == :special_inside_parentheses
               left_parenthesis.column + 1
@@ -78,7 +102,7 @@ module Rubocop
               left_brace.source_line =~ /\S/
             end
 
-          base_column + IndentationWidth::CORRECT_INDENTATION
+          base_column + IndentationWidth::CORRECT_INDENTATION + offset
         end
 
         # Returns the description of what the correct indentation is based on.
@@ -92,7 +116,7 @@ module Rubocop
 
         # Returns the "unexpected column", which is the column that would be
         # correct if the configuration was changed.
-        def unexpected_column(left_brace, left_parenthesis)
+        def unexpected_column(left_brace, left_parenthesis, offset)
           # Set a crazy value by default, indicating that there's no other
           # configuration that can be chosen to make the used indentation
           # accepted.
@@ -106,7 +130,8 @@ module Rubocop
                                      end
           end
 
-          unexpected_base_column + IndentationWidth::CORRECT_INDENTATION
+          unexpected_base_column + IndentationWidth::CORRECT_INDENTATION +
+            offset
         end
 
         def message(base_description)